refactor: restructure application folders and standardize component naming
